fix(suspense): 배포 suspense 오류 수정
