+++ .env.local
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_API_TOKEN = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlYjA5ZTgyOS02MDYxLTQwMDgtYTRjZC1iMzc0NWU5ODhmMjkiLCJpYXQiOjE3MzUwNTI4NzN9.kXzNSMPekR6xyW6lH_H3iw7MBNxyk3z0nZq_g-TPqgU'
